{{cookiecutter.slug}}/requirements.txt
stepup==2.0.2
stepup-reprep==2.1.0
pyyaml
pre-commit
